requirements.txt
python-dotenv==1
praw>=7.4.0
